[FP-73] Refactor: 타입 폴더 위치 변경
