+++ src/components/common/CardResource/cardResource.types.ts
