+++ src/components/pages/myPage/Schedule/reservation.types.ts
